requirements.txt
requests>=2.28.0
bs4>=0.0.1
pandas>=1.4.2
mysql-connector-python>=8.0.29